cleans up section block
